// frontend/src/api/api8082/models/BookWithOwnerDto.ts
/* tslint:disable */
/* eslint-disable */
/**
 * Authorization server
 * No description provided (generated by Openapi Generator https://github.com/openapitools/openapi-generator)
 *
 * The version of the OpenAPI document: v1
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

import { mapValues } from '../runtime';
/**
 * 
 * @export
 * @interface BookWithOwnerDto
 */
export interface BookWithOwnerDto {
    /**
     * 
     * @type {string}
     * @memberof BookWithOwnerDto
     */
    title?: string;
    /**
     * 
     * @type {string}
     * @memberof BookWithOwnerDto
     */
    author?: string;
    /**
     * 
     * @type {string}
     * @memberof BookWithOwnerDto
     */
    ownerEmail?: string;
    /**
     * 
     * @type {string}
     * @memberof BookWithOwnerDto
     */
    ownerUsername?: string;
}

/**
 * Check if a given object implements the BookWithOwnerDto interface.
 */
export function instanceOfBookWithOwnerDto(value: object): value is BookWithOwnerDto {
    return true;
}

export function BookWithOwnerDtoFromJSON(json: any): BookWithOwnerDto {
    return BookWithOwnerDtoFromJSONTyped(json, false);
}

export function BookWithOwnerDtoFromJSONTyped(json: any, ignoreDiscriminator: boolean): BookWithOwnerDto {
    if (json == null) {
        return json;
    }
    return {
        
        'title': json['title'] == null ? undefined : json['title'],
        'author': json['author'] == null ? undefined : json['author'],
        'ownerEmail': json['ownerEmail'] == null ? undefined : json['ownerEmail'],
        'ownerUsername': json['ownerUsername'] == null ? undefined : json['ownerUsername'],
    };
}

export function BookWithOwnerDtoToJSON(json: any): BookWithOwnerDto {
    return BookWithOwnerDtoToJSONTyped(json, false);
}

export function BookWithOwnerDtoToJSONTyped(value?: BookWithOwnerDto | null, ignoreDiscriminator: boolean = false): any {
    if (value == null) {
        return value;
    }

    return {
        
        'title': value['title'],
        'author': value['author'],
        'ownerEmail': value['ownerEmail'],
        'ownerUsername': value['ownerUsername'],
    };
}