// frontend/src/api/api8082/models/OfferDto.ts
/* tslint:disable */
/* eslint-disable */
/**
 * Authorization server
 * No description provided (generated by Openapi Generator https://github.com/openapitools/openapi-generator)
 *
 * The version of the OpenAPI document: v1
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

import { mapValues } from '../runtime';
/**
 * 
 * @export
 * @interface OfferDto
 */
export interface OfferDto {
    /**
     * 
     * @type {string}
     * @memberof OfferDto
     */
    id?: string;
    /**
     * 
     * @type {string}
     * @memberof OfferDto
     */
    senderEmail?: string;
    /**
     * 
     * @type {string}
     * @memberof OfferDto
     */
    receiverEmail?: string;
    /**
     * 
     * @type {Array<string>}
     * @memberof OfferDto
     */
    offeredBookTitles?: Array<string>;
    /**
     * 
     * @type {Array<string>}
     * @memberof OfferDto
     */
    requestedBookTitles?: Array<string>;
    /**
     * 
     * @type {string}
     * @memberof OfferDto
     */
    status?: string;
}

/**
 * Check if a given object implements the OfferDto interface.
 */
export function instanceOfOfferDto(value: object): value is OfferDto {
    return true;
}

export function OfferDtoFromJSON(json: any): OfferDto {
    return OfferDtoFromJSONTyped(json, false);
}

export function OfferDtoFromJSONTyped(json: any, ignoreDiscriminator: boolean): OfferDto {
    if (json == null) {
        return json;
    }
    return {
        
        'id': json['id'] == null ? undefined : json['id'],
        'senderEmail': json['senderEmail'] == null ? undefined : json['senderEmail'],
        'receiverEmail': json['receiverEmail'] == null ? undefined : json['receiverEmail'],
        'offeredBookTitles': json['offeredBookTitles'] == null ? undefined : json['offeredBookTitles'],
        'requestedBookTitles': json['requestedBookTitles'] == null ? undefined : json['requestedBookTitles'],
        'status': json['status'] == null ? undefined : json['status'],
    };
}

export function OfferDtoToJSON(json: any): OfferDto {
    return OfferDtoToJSONTyped(json, false);
}

export function OfferDtoToJSONTyped(value?: OfferDto | null, ignoreDiscriminator: boolean = false): any {
    if (value == null) {
        return value;
    }

    return {
        
        'id': value['id'],
        'senderEmail': value['senderEmail'],
        'receiverEmail': value['receiverEmail'],
        'offeredBookTitles': value['offeredBookTitles'],
        'requestedBookTitles': value['requestedBookTitles'],
        'status': value['status'],
    };
}